+++ otx/algorithms/segmentation/adapters/mmseg/models/losses/__init__.py
@@ -16,5 +16,6 @@
 
 from .cross_entropy_loss_with_ignore import CrossEntropyLossWithIgnore
 from .detcon_loss import DetConLoss
+from .pixel_prototype_ce_loss import PixelPrototypeCELoss
 
-__all__ = ["DetConLoss", "CrossEntropyLossWithIgnore"]
+__all__ = ["DetConLoss", "CrossEntropyLossWithIgnore", "PixelPrototypeCELoss"]