+++ otx/algorithms/segmentation/configs/ham_segnext_t/model.py
@@ -40,4 +40,5 @@ model = dict(
     ),
 )
 
+optimizer = dict(paramwise_cfg=dict(custom_keys={"pos_block": dict(decay_mult=0.0), "norm": dict(decay_mult=0.0)}))
 load_from = "https://download.openmmlab.com/mmsegmentation/v0.5/pretrain/segnext/mscan_t_20230227-119e8c9f.pth"