+++ tests/unit/cli/utils/test_hpo.py
@@ -476,6 +476,16 @@ class TestHpoRunner:
 
 
 class TestTrainer:
+    @pytest.fixture(autouse=True)
+    def setup(self, tmp_dir):
+        self.weight_format = "epoch_{}.pth"
+        self.hpo_workdir = Path(tmp_dir) / "hpo_dir"
+
+    @pytest.fixture
+    def tmp_dir(self):
+        with TemporaryDirectory() as tmp_dir:
+            yield tmp_dir
+
     @e2e_pytest_unit
     def test_init(self, mocker, cls_template_path):
         Trainer(
@@ -489,48 +499,81 @@ class TestTrainer:
             metric="fake",
         )
 
+    @pytest.fixture
+    def mock_task(self, mocker, tmp_dir):
+        fake_project_path = Path(tmp_dir) / "fake_proejct"
+        fake_project_path.mkdir(parents=True)
+        for i in range(1, 5):
+            (fake_project_path / self.weight_format.format(i)).write_text("fake")
+
+        mock_get_train_task = mocker.patch.object(TaskEnvironmentManager, "get_train_task")
+        mock_task = mocker.MagicMock()
+        mock_task.project_path = str(fake_project_path)
+        mock_get_train_task.return_value = mock_task
+
+        return mock_task
+
     @e2e_pytest_unit
-    def test_run(self, mocker, cls_template_path):
-        with TemporaryDirectory() as tmp_dir:
-            # prepare
-            trial_id = "1"
-            weight_format = "epoch_{}.pth"
-            hpo_workdir = Path(tmp_dir) / "hpo_dir"
-            fake_project_path = Path(tmp_dir) / "fake_proejct"
-            fake_project_path.mkdir(parents=True)
-            for i in range(1, 5):
-                (fake_project_path / weight_format.format(i)).write_text("fake")
+    def test_run(self, mocker, cls_template_path, mock_task, tmp_dir):
+        # prepare
+        trial_id = "1"
+        mock_report_func = mocker.MagicMock()
 
-            mock_get_train_task = mocker.patch.object(TaskEnvironmentManager, "get_train_task")
-            mock_task = mocker.MagicMock()
-            mock_task.project_path = str(fake_project_path)
-            mock_get_train_task.return_value = mock_task
+        mocker.patch("otx.cli.utils.hpo.get_dataset_adapter")
+        mocker.patch("otx.cli.utils.hpo.HpoDataset")
 
-            mock_report_func = mocker.MagicMock()
+        # run
+        trainer = Trainer(
+            hp_config={"configuration": {"iterations": 10}, "id": trial_id},
+            report_func=mock_report_func,
+            model_template=find_and_parse_model_template(cls_template_path),
+            data_roots=mocker.MagicMock(),
+            task_type=TaskType.CLASSIFICATION,
+            hpo_workdir=self.hpo_workdir,
+            initial_weight_name="fake",
+            metric="fake",
+        )
+        trainer.run()
 
-            mocker.patch("otx.cli.utils.hpo.get_dataset_adapter")
-            mocker.patch("otx.cli.utils.hpo.HpoDataset")
-
-            # run
-            trainer = Trainer(
-                hp_config={"configuration": {"iterations": 10}, "id": trial_id},
-                report_func=mock_report_func,
-                model_template=find_and_parse_model_template(cls_template_path),
-                data_roots=mocker.MagicMock(),
-                task_type=TaskType.CLASSIFICATION,
-                hpo_workdir=hpo_workdir,
-                initial_weight_name="fake",
-                metric="fake",
-            )
-            trainer.run()
-
-            # check
-            mock_report_func.assert_called_once_with(0, 0, done=True)  # finilize report
-            assert hpo_workdir.exists()  # make a directory to copy weight
-            for i in range(1, 5):  # check model weights are copied
-                assert (hpo_workdir / "weight" / trial_id / weight_format.format(i)).exists()
+        # check
+        mock_report_func.assert_called_once_with(0, 0, done=True)  # finilize report
+        assert self.hpo_workdir.exists()  # make a directory to copy weight
+        for i in range(1, 5):  # check model weights are copied
+            assert (self.hpo_workdir / "weight" / trial_id / self.weight_format.format(i)).exists()
 
         mock_task.train.assert_called()  # check task.train() is called
+
+    @e2e_pytest_unit
+    def test_run_trial_already_done(self, mocker, cls_template_path, mock_task, tmp_dir):
+        """Test a case where trial to run already training given epoch."""
+        # prepare
+        trial_id = "1"
+        epoch_to_run = 10
+        weight_dir = self.hpo_workdir / "weight" / trial_id
+        # prepare a weight trained more than given epoch
+        weight_dir.mkdir(parents=True)
+        (weight_dir / self.weight_format.format(epoch_to_run + 1)).touch()
+        mock_report_func = mocker.MagicMock()
+
+        mocker.patch("otx.cli.utils.hpo.get_dataset_adapter")
+        mocker.patch("otx.cli.utils.hpo.HpoDataset")
+
+        # run
+        trainer = Trainer(
+            hp_config={"configuration": {"iterations": epoch_to_run}, "id": trial_id},
+            report_func=mock_report_func,
+            model_template=find_and_parse_model_template(cls_template_path),
+            data_roots=mocker.MagicMock(),
+            task_type=TaskType.CLASSIFICATION,
+            hpo_workdir=self.hpo_workdir,
+            initial_weight_name="fake",
+            metric="fake",
+        )
+        trainer.run()
+
+        # check
+        mock_report_func.assert_called_once_with(0, 0, done=True)  # finilize report
+        mock_task.train.assert_not_called()  # check task.train() is called
 
 
 class TestHpoCallback:
@@ -617,6 +660,11 @@ def test_run_hpo(mocker, mock_environment):
         output = Path(tmp_dir) / "fake"
         mock_get_best_hpo_weight = mocker.patch("otx.cli.utils.hpo.get_best_hpo_weight")
         mock_get_best_hpo_weight.return_value = "mock_best_weight_path"
+        hpo_weight_dir = output / "hpo" / "weight"
+        for i in range(3):
+            trial_weight_dir = hpo_weight_dir / str(i)
+            trial_weight_dir.mkdir(parents=True)
+            (trial_weight_dir / "fake.pth").touch()
 
         def mock_run_hpo(*args, **kwargs):
             return {"config": {"a.b": 1, "c.d": 2}, "id": "1"}
@@ -643,6 +691,7 @@ def test_run_hpo(mocker, mock_environment):
         assert env_hp.a.b == 1
         assert env_hp.c.d == 2
         assert environment.model == "mock_best_weight_path"  # check that best model weight is used
+        assert not list(hpo_weight_dir.rglob("*.pth"))  # check unused weight is removed
 
 
 @e2e_pytest_unit